feat: add default cash account flow

+++ packages/backend/convex/transactions.ts
@@ -1,5 +1,5 @@
 import schema from "./schema";
-import {  aggregateMonthlyBalanceByUser, aggregateMonthlyTransactionsByUser } from "./schema";
+import { aggregateMonthlyTransactionsByUser } from "./schema";
 import { mutation, query } from "./_generated/server";
 import { v } from "convex/values";
 import { authComponent } from "./auth";
@@ -27,24 +27,24 @@ export const listMonthlyTransactions = query({
   },
 });
 
-export const getMonthlyBalance = query({
-  args: { monthStart: v.string(), monthEnd: v.string() },
-  handler: async (ctx, { monthStart, monthEnd }) => {
-    const user = await authComponent.getAuthUser(ctx);
+// export const getMonthlyBalance = query({
+//   args: { monthStart: v.string(), monthEnd: v.string() },
+//   handler: async (ctx, { monthStart, monthEnd }) => {
+//     const user = await authComponent.getAuthUser(ctx);
 
-    if (!user) {
-      return null;
-    }
+//     if (!user) {
+//       return null;
+//     }
 
-    return await aggregateMonthlyBalanceByUser.sum(ctx, {
-      bounds: {
-        lower: {key: monthStart, inclusive: true},
-        upper: {key: monthEnd, inclusive: true}
-      },
-      namespace: user._id
-    })
-  }
-})
+//     return await aggregateMonthlyBalanceByUser.sum(ctx, {
+//       bounds: {
+//         lower: {key: monthStart, inclusive: true},
+//         upper: {key: monthEnd, inclusive: true}
+//       },
+//       namespace: user._id
+//     })
+//   }
+// })
 
 export const createTransaction = mutation({
   args: {
@@ -52,11 +52,18 @@ export const createTransaction = mutation({
     description: schema.tables.transactions.validator.fields.description,
     category: schema.tables.transactions.validator.fields.category,
     date: schema.tables.transactions.validator.fields.date,
+    accountId: schema.tables.transactions.validator.fields.accountId,
   },
   handler: async (ctx, params) => {
 
     if (params.amount === 0) {
       throw new Error("Amount must be greater than zero");
+    }
+
+    const account = await ctx.db.get(params.accountId)
+
+    if (account === null) {
+      throw new Error("Account not found");
     }
 
     const user = await authComponent.getAuthUser(ctx);
@@ -66,8 +73,12 @@ export const createTransaction = mutation({
       amount: params.category === "income" ? Math.abs(params.amount) : -Math.abs(params.amount),
       userId: user._id
     })
+
+    await ctx.db.patch("accounts", params.accountId, {
+      balance: account.balance + params.amount
+    })
+
     const doc = await ctx.db.get(id)
-    await  aggregateMonthlyBalanceByUser.insert(ctx, doc!)
     await aggregateMonthlyTransactionsByUser.insert(ctx, doc!)
     return id;
   }
@@ -76,10 +87,26 @@ export const createTransaction = mutation({
 export const deleteTransaction = mutation({
   args: { id: v.id("transactions") },
   handler: async (ctx, {id}) => {
-    const doc = await ctx.db.get(id)
-    await aggregateMonthlyBalanceByUser.deleteIfExists(ctx, doc!)
-    await aggregateMonthlyTransactionsByUser.deleteIfExists(ctx, doc!)
+
+    const transaction = await ctx.db.get(id)
+
+    if (transaction === null) {
+      throw new Error("Transaction not found");
+    }
+
     await ctx.db.delete("transactions", id);
+    await aggregateMonthlyTransactionsByUser.deleteIfExists(ctx, transaction)
+
+    const account = await ctx.db.get(transaction.accountId)
+
+    if (account === null) {
+      throw new Error("Account not found");
+    }
+
+    await ctx.db.patch("accounts", transaction.accountId, {
+      balance: account.balance - transaction.amount
+    })
+
     return id;
   }
 })
